Include all files in the IPFS output directory
This commit changes the behavior when a UnixFS directory is created
containing all the files in artifacts so that it includes the latest
version (by git committer time) of each file (by filename) in each
artifact.

This commit also changes the API to add a third mode: `upload`. This
ensures that uploading files always traverses the entire history.